Expense Tracker/requirements.txt
django
django-filter
djangorestframework
drf-yasg
gunicorn
whitenoise
psycopg2-binary
python-dotenv